pyproject.toml
[project]
name = "raffle-backoffice"
version = "0.1.0"
description = "Deterministic raffle ticket chatbot backoffice using Meta WhatsApp Cloud API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.15.0",
    "django-jazzmin>=3.0.0",
    "django-filter>=24.0",
    "drf-spectacular>=0.27.0",
    "requests>=2.32.0",
    "python-decouple>=3.8",
]
